Add a better onbuild image for Go programs.
It's based on alpine (2M base image) instead of ubuntu, and deletes
all the packages and source after building, so you don't end up
with 130M of Go toolchain when you're done.

This is really just a workaround for docker not having some kind
of "flatten" dockerfile, so you constantly carry garbage around
in your images.